+++ frontend/src/contexts/NavigationBarProvider.tsx
