+++ frontend/src/hooks/useNavigationBarDispatchContext.ts
@@ -1,5 +1,5 @@
 import { useContext } from 'react';
-import { NavigationBarDispatchContext } from '@/context/NavigationBarProvider';
+import { NavigationBarDispatchContext } from '@/contexts/NavigationBarProvider';
 
 function useNavigationBarDispatchContext() {
   const context = useContext(NavigationBarDispatchContext);
